% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocWorkshopSubmit.R
\name{BiocWorkshopSubmit}
\alias{BiocWorkshopSubmit}
\title{Prepare a Bioconductor Workshop Submission}
\usage{
BiocWorkshopSubmit(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{shinyApp}}}
}
\description{
Prepare a Bioconductor Workshop Submission
}
